+++ noodles-cram/src/data_container/compression_header/preservation_map/substitution_matrix/builder.rs
@@ -55,7 +55,7 @@ mod tests {
         let bases: Sequence = "TCTGGCGTGT".parse()?;
 
         let record = Record::builder()
-            .set_alignment_start(sam::record::Position::try_from(1)?)
+            .set_alignment_start(Position::try_from(1)?)
             .set_read_length(bases.len())
             .set_bases(bases)
             .add_feature(Feature::Substitution(Position::try_from(1)?, 2)) // A => T